refactor: rename project page to works and move some content about page to it
